refactor(typecheck-worker): move type_extractor tests to separate file
Extract the #[cfg(test)] module from type_extractor.rs into
type_extractor_tests.rs (included via #[path]) to bring the file
from 620 lines to 397 lines, satisfying the 600-line CI cap.
No logic changes; all 43 unit + 35 integration tests pass.
